refactor(expo): clean up native prebuilt views
