+++ packages/expo/ios/ClerkExpoModule.swift
@@ -12,9 +12,9 @@ public var clerkViewFactory: ClerkViewFactoryProtocol?
 // Protocol that the app target implements to provide Clerk views
 public protocol ClerkViewFactoryProtocol {
   // Inline rendering — returns UIViewController to preserve SwiftUI lifecycle
-  func createAuthView(mode: String, dismissable: Bool, onEvent: @escaping (String, [String: Any]) -> Void) -> UIViewController?
-  func createUserProfileView(dismissable: Bool, onEvent: @escaping (String, [String: Any]) -> Void) -> UIViewController?
-  func createUserButton(onEvent: @escaping (String, [String: Any]) -> Void) -> UIViewController?
+  func createAuthView(mode: String, dismissable: Bool, onEvent: @escaping (ClerkNativeViewEvent, [String: Any]) -> Void) -> UIViewController?
+  func createUserProfileView(dismissable: Bool, onEvent: @escaping (ClerkNativeViewEvent, [String: Any]) -> Void) -> UIViewController?
+  func createUserButton(onEvent: @escaping (ClerkNativeViewEvent, [String: Any]) -> Void) -> UIViewController?
 
   // SDK operations
   func configure(publishableKey: String, bearerToken: String?) async throws
@@ -54,10 +54,10 @@ class ClerkExpoModule: RCTEventEmitter {
 
   /// Emits an onAuthStateChange event to JS from anywhere in the native layer.
   /// Used by native views to notify ClerkProvider of auth state changes.
-  static func emitAuthStateChange(type: String, sessionId: String?) {
+  static func emitAuthStateChange(type: ClerkNativeAuthStateEvent, sessionId: String?) {
     guard _hasListeners, let instance = sharedInstance else { return }
     instance.sendEvent(withName: "onAuthStateChange", body: [
-      "type": type,
+      "type": type.rawValue,
       "sessionId": sessionId as Any,
     ])
   }
@@ -127,300 +127,5 @@ class ClerkExpoModule: RCTEventEmitter {
         reject("E_SIGN_OUT_FAILED", error.localizedDescription, error)
       }
     }
-  }
-}
-
-// MARK: - Inline View: ClerkUserButtonNativeView
-
-public class ClerkUserButtonNativeView: UIView {
-  private var hostingController: UIViewController?
-  private var hasInitialized: Bool = false
-
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  override public func didMoveToWindow() {
-    super.didMoveToWindow()
-    if window != nil && !hasInitialized {
-      hasInitialized = true
-      updateView()
-    }
-  }
-
-  private func updateView() {
-    detachHostingController()
-
-    guard let factory = clerkViewFactory else { return }
-
-    guard let returnedController = factory.createUserButton(
-      onEvent: { eventName, data in
-        if eventName == "signedOut" {
-          let sessionId = data["sessionId"] as? String
-          ClerkExpoModule.emitAuthStateChange(type: "signedOut", sessionId: sessionId)
-        }
-      }
-    ) else { return }
-
-    attachHostingController(returnedController)
-  }
-
-  private func attachHostingController(_ controller: UIViewController) {
-    controller.view.frame = bounds
-    controller.view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-    controller.view.backgroundColor = .clear
-
-    if let parentVC = findViewController() {
-      parentVC.addChild(controller)
-      addSubview(controller.view)
-      controller.didMove(toParent: parentVC)
-    } else {
-      addSubview(controller.view)
-    }
-
-    hostingController = controller
-  }
-
-  private func detachHostingController() {
-    guard let controller = hostingController else { return }
-    controller.willMove(toParent: nil)
-    controller.view.removeFromSuperview()
-    controller.removeFromParent()
-    hostingController = nil
-  }
-
-  private func findViewController() -> UIViewController? {
-    var responder: UIResponder? = self
-    while let nextResponder = responder?.next {
-      if let vc = nextResponder as? UIViewController {
-        return vc
-      }
-      responder = nextResponder
-    }
-    return nil
-  }
-
-  override public func layoutSubviews() {
-    super.layoutSubviews()
-    hostingController?.view.frame = bounds
-  }
-}
-
-// MARK: - Inline View: ClerkAuthNativeView
-
-public class ClerkAuthNativeView: UIView {
-  private var hostingController: UIViewController?
-  private var currentMode: String = "signInOrUp"
-  private var currentDismissable: Bool = false
-  private var hasInitialized: Bool = false
-  private var didCompleteAuthentication: Bool = false
-  private var dismissalEventSent: Bool = false
-
-  @objc var onAuthEvent: RCTBubblingEventBlock?
-
-  @objc var mode: NSString? {
-    didSet {
-      let newMode = (mode as String?) ?? "signInOrUp"
-      guard newMode != currentMode else { return }
-      currentMode = newMode
-      if hasInitialized { updateView() }
-    }
-  }
-
-  @objc var isDismissable: NSNumber? {
-    didSet {
-      let newDismissable = isDismissable?.boolValue ?? false
-      guard newDismissable != currentDismissable else { return }
-      currentDismissable = newDismissable
-      if hasInitialized { updateView() }
-    }
-  }
-
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  override public func didMoveToWindow() {
-    super.didMoveToWindow()
-    if window != nil && !hasInitialized {
-      hasInitialized = true
-      updateView()
-    } else if window == nil && hasInitialized && currentDismissable && !didCompleteAuthentication && !dismissalEventSent {
-      dismissalEventSent = true
-      sendAuthEvent(type: "dismissed")
-    }
-  }
-
-  private func sendAuthEvent(type: String) {
-    onAuthEvent?(["type": type])
-  }
-
-  private func updateView() {
-    detachHostingController()
-
-    guard let factory = clerkViewFactory else { return }
-
-    guard let returnedController = factory.createAuthView(
-      mode: currentMode,
-      dismissable: currentDismissable,
-      onEvent: { [weak self] eventName, data in
-        let didCompleteAuthentication = eventName == "signInCompleted" || eventName == "signUpCompleted"
-
-        if didCompleteAuthentication {
-          self?.didCompleteAuthentication = true
-          let sessionId = data["sessionId"] as? String
-          ClerkExpoModule.emitAuthStateChange(type: "signedIn", sessionId: sessionId)
-        }
-      }
-    ) else { return }
-
-    attachHostingController(returnedController)
-  }
-
-  private func attachHostingController(_ controller: UIViewController) {
-    controller.view.frame = bounds
-    controller.view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-
-    if let parentVC = findViewController() {
-      parentVC.addChild(controller)
-      addSubview(controller.view)
-      controller.didMove(toParent: parentVC)
-    } else {
-      addSubview(controller.view)
-    }
-
-    hostingController = controller
-  }
-
-  private func detachHostingController() {
-    guard let controller = hostingController else { return }
-    controller.willMove(toParent: nil)
-    controller.view.removeFromSuperview()
-    controller.removeFromParent()
-    hostingController = nil
-  }
-
-  private func findViewController() -> UIViewController? {
-    var responder: UIResponder? = self
-    while let nextResponder = responder?.next {
-      if let vc = nextResponder as? UIViewController {
-        return vc
-      }
-      responder = nextResponder
-    }
-    return nil
-  }
-
-  override public func layoutSubviews() {
-    super.layoutSubviews()
-    hostingController?.view.frame = bounds
-  }
-}
-
-// MARK: - Inline View: ClerkUserProfileNativeView
-
-public class ClerkUserProfileNativeView: UIView {
-  private var hostingController: UIViewController?
-  private var currentDismissable: Bool = false
-  private var hasInitialized: Bool = false
-  private var didSignOut = false
-  private var dismissalEventSent = false
-
-  @objc var onProfileEvent: RCTBubblingEventBlock?
-
-  @objc var isDismissable: NSNumber? {
-    didSet {
-      currentDismissable = isDismissable?.boolValue ?? false
-      if hasInitialized { updateView() }
-    }
-  }
-
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  override public func didMoveToWindow() {
-    super.didMoveToWindow()
-    if window != nil && !hasInitialized {
-      hasInitialized = true
-      updateView()
-    } else if window == nil && hasInitialized && currentDismissable && !didSignOut && !dismissalEventSent {
-      dismissalEventSent = true
-      sendProfileEvent(type: "dismissed")
-    }
-  }
-
-  private func sendProfileEvent(type: String) {
-    onProfileEvent?(["type": type])
-  }
-
-  private func updateView() {
-    detachHostingController()
-
-    guard let factory = clerkViewFactory else { return }
-
-    guard let returnedController = factory.createUserProfileView(
-      dismissable: currentDismissable,
-      onEvent: { [weak self] eventName, data in
-        if eventName == "signedOut" {
-          self?.didSignOut = true
-          let sessionId = data["sessionId"] as? String
-          ClerkExpoModule.emitAuthStateChange(type: "signedOut", sessionId: sessionId)
-        }
-      }
-    ) else { return }
-
-    attachHostingController(returnedController)
-  }
-
-  private func attachHostingController(_ controller: UIViewController) {
-    controller.view.frame = bounds
-    controller.view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-
-    if let parentVC = findViewController() {
-      parentVC.addChild(controller)
-      addSubview(controller.view)
-      controller.didMove(toParent: parentVC)
-    } else {
-      addSubview(controller.view)
-    }
-
-    hostingController = controller
-  }
-
-  private func detachHostingController() {
-    guard let controller = hostingController else { return }
-    controller.willMove(toParent: nil)
-    controller.view.removeFromSuperview()
-    controller.removeFromParent()
-    hostingController = nil
-  }
-
-  private func findViewController() -> UIViewController? {
-    var responder: UIResponder? = self
-    while let nextResponder = responder?.next {
-      if let vc = nextResponder as? UIViewController {
-        return vc
-      }
-      responder = nextResponder
-    }
-    return nil
-  }
-
-  override public func layoutSubviews() {
-    super.layoutSubviews()
-    hostingController?.view.frame = bounds
   }
 }
